pyproject.toml
[tool.poetry]
name = "wordgame-bot"
version = "0.1.0"
description = ""
authors = ["Harry Williams <1998.harrywilliams@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.10"
freezegun = "^1.1.0"
discord = "^1.7.3"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
coverage = "^6.3.2"
pytest = "7.0.1"
pytest-asyncio = "^0.18.2"
python-dotenv = "0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
